Update lexer templates for Go with some micro optimizations.

+++ tm-go/parsers/json/parser_tables.go
@@ -19,10 +19,10 @@ var symbolStr = [...]string{
 }
 
 func (n Symbol) String() string {
-    if n < Symbol(terminalEnd) {
-      return Token(n).String()
-    }
-    i := int(n) - int(terminalEnd)
+	if n < Symbol(terminalEnd) {
+		return Token(n).String()
+	}
+	i := int(n) - int(terminalEnd)
 	if i < len(symbolStr) {
 		return symbolStr[i]
 	}
@@ -45,21 +45,21 @@ var tmGoto = []int32{
 	41, 41, 42, 43, 46, 50, 52, 53, 57, 58, 59,
 }
 
-var tmFrom = []int32{
+var tmFrom = []int8{
 	35, 0, 2, 19, 22, 1, 15, 0, 2, 19, 22, 18, 13, 15, 17, 0,
 	1, 2, 19, 21, 22, 0, 2, 19, 22, 0, 2, 19, 22, 0, 2, 19,
 	22, 0, 2, 19, 22, 0, 2, 22, 19, 0, 19, 0, 2, 22, 0, 2,
 	19, 22, 1, 21, 1, 0, 2, 19, 22, 2, 2,
 }
 
-var tmTo = []int32{
+var tmTo = []int8{
 	36, 1, 1, 1, 1, 12, 20, 2, 2, 2, 2, 23, 19, 21, 22, 3,
 	13, 3, 24, 13, 3, 4, 4, 25, 4, 5, 5, 26, 5, 6, 6, 27,
 	6, 7, 7, 28, 7, 8, 8, 8, 29, 35, 30, 9, 16, 34, 10, 10,
 	31, 10, 14, 33, 15, 11, 11, 32, 11, 17, 18,
 }
 
-var tmRuleLen = []int32{
+var tmRuleLen = []int8{
 	1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1, 1,
 	1, 3, 2, 3, 1, 3, 3, 1, 3, 1, 0,
 }
